Allow 'python setup.py install' to work correctly for either Python 2 or 3.
