+++ setupbase.py
@@ -245,28 +245,31 @@ def make_man_update_target(manpage):
 # Find scripts
 #---------------------------------------------------------------------------
 
-def find_scripts(entry_points=False):
+def find_scripts(entry_points=False, suffix=''):
     """Find IPython's scripts.
     
     if entry_points is True:
         return setuptools entry_point-style definitions
     else:
         return file paths of plain scripts [default]
+    
+    suffix is appended to script names if entry_points is True, so that the
+    Python 3 scripts get named "ipython3" etc.
     """
     if entry_points:
-        console_scripts = [
-            'ipython = IPython.frontend.terminal.ipapp:launch_new_instance',
-            'pycolor = IPython.utils.PyColorize:main',
-            'ipcontroller = IPython.parallel.apps.ipcontrollerapp:launch_new_instance',
-            'ipengine = IPython.parallel.apps.ipengineapp:launch_new_instance',
-            'iplogger = IPython.parallel.apps.iploggerapp:launch_new_instance',
-            'ipcluster = IPython.parallel.apps.ipclusterapp:launch_new_instance',
-            'iptest = IPython.testing.iptest:main',
-            'irunner = IPython.lib.irunner:main'
-        ]
-        gui_scripts = [
-            'ipython-qtconsole = IPython.frontend.qt.console.qtconsoleapp:main',
-        ]
+        console_scripts = [s % suffix for s in [
+            'ipython%s = IPython.frontend.terminal.ipapp:launch_new_instance',
+            'pycolor%s = IPython.utils.PyColorize:main',
+            'ipcontroller%s = IPython.parallel.apps.ipcontrollerapp:launch_new_instance',
+            'ipengine%s = IPython.parallel.apps.ipengineapp:launch_new_instance',
+            'iplogger%s = IPython.parallel.apps.iploggerapp:launch_new_instance',
+            'ipcluster%s = IPython.parallel.apps.ipclusterapp:launch_new_instance',
+            'iptest%s = IPython.testing.iptest:main',
+            'irunner%s = IPython.lib.irunner:main'
+        ]]
+        gui_scripts = [s % suffix for s in [
+            'ipython%s-qtconsole = IPython.frontend.qt.console.qtconsoleapp:main',
+        ]]
         scripts = dict(console_scripts=console_scripts, gui_scripts=gui_scripts)
     else:
         parallel_scripts = pjoin('IPython','parallel','scripts')
